+++ src/axom/inlet/tests/inlet_object.cpp
@@ -47,7 +47,7 @@ struct Foo
 template <>
 struct FromInlet<Foo>
 {
-  Foo operator()(const axom::inlet::Table& base)
+  Foo operator()(const axom::inlet::Container& base)
   {
     Foo f {base["bar"], base["baz"]};
     return f;
@@ -85,10 +85,10 @@ TYPED_TEST(inlet_object, simple_array_of_struct_by_value)
   DataStore ds;
   Inlet inlet = createBasicInlet<TypeParam>(&ds, testString);
 
-  auto& arr_table = inlet.addStructArray("foo");
+  auto& arr_container = inlet.addStructArray("foo");
 
-  arr_table.addBool("bar", "bar's description");
-  arr_table.addBool("baz", "baz's description");
+  arr_container.addBool("bar", "bar's description");
+  arr_container.addBool("baz", "baz's description");
   std::unordered_map<int, Foo> expected_foos = {{0, {true, false}},
                                                 {1, {false, true}}};
   auto foos = inlet["foo"].get<std::unordered_map<int, Foo>>();
@@ -103,10 +103,10 @@ TYPED_TEST(inlet_object, simple_array_of_struct_implicit_idx)
   DataStore ds;
   Inlet inlet = createBasicInlet<TypeParam>(&ds, testString);
 
-  auto& arr_table = inlet.addStructArray("foo");
+  auto& arr_container = inlet.addStructArray("foo");
 
-  arr_table.addBool("bar", "bar's description");
-  arr_table.addBool("baz", "baz's description");
+  arr_container.addBool("bar", "bar's description");
+  arr_container.addBool("baz", "baz's description");
   // Lua is 1-indexed
   const int base_idx = TypeParam::baseIndex;
   std::unordered_map<int, Foo> expected_foos = {{base_idx, {true, false}},
@@ -123,10 +123,10 @@ TYPED_TEST(inlet_object, simple_array_of_struct_verify_optional)
   DataStore ds;
   Inlet inlet = createBasicInlet<TypeParam>(&ds, testString);
 
-  auto& arr_table = inlet.addStructArray("foo");
+  auto& arr_container = inlet.addStructArray("foo");
 
-  arr_table.addBool("bar", "bar's description").required(true);
-  arr_table.addBool("baz", "baz's description").required(false);
+  arr_container.addBool("bar", "bar's description").required(true);
+  arr_container.addBool("baz", "baz's description").required(false);
 
   EXPECT_TRUE(inlet.verify());
 }
@@ -139,10 +139,10 @@ TYPED_TEST(inlet_object, simple_array_of_struct_verify_reqd)
   DataStore ds;
   Inlet inlet = createBasicInlet<TypeParam>(&ds, testString);
 
-  auto& arr_table = inlet.addStructArray("foo");
+  auto& arr_container = inlet.addStructArray("foo");
 
-  arr_table.addBool("bar", "bar's description").required(true);
-  arr_table.addBool("baz", "baz's description").required(true);
+  arr_container.addBool("bar", "bar's description").required(true);
+  arr_container.addBool("baz", "baz's description").required(true);
 
   EXPECT_FALSE(inlet.verify());
 }
@@ -153,11 +153,11 @@ TYPED_TEST(inlet_object, simple_array_of_struct_verify_empty_pass)
   DataStore ds;
   Inlet inlet = createBasicInlet<TypeParam>(&ds, testString);
 
-  auto& arr_table = inlet.addStructArray("foo");
+  auto& arr_container = inlet.addStructArray("foo");
   // Even though these are required, the input file should still
   // "verify" as the array is empty
-  arr_table.addBool("bar", "bar's description").required(true);
-  arr_table.addBool("baz", "baz's description").required(true);
+  arr_container.addBool("bar", "bar's description").required(true);
+  arr_container.addBool("baz", "baz's description").required(true);
 
   EXPECT_TRUE(inlet.verify());
 }
@@ -170,9 +170,9 @@ TYPED_TEST(inlet_object, simple_array_of_struct_verify_empty_fail)
 
   // Verification should fail because the array is empty
   // and was required
-  auto& arr_table = inlet.addStructArray("foo").required(true);
-  arr_table.addBool("bar", "bar's description").required(true);
-  arr_table.addBool("baz", "baz's description").required(true);
+  auto& arr_container = inlet.addStructArray("foo").required(true);
+  arr_container.addBool("bar", "bar's description").required(true);
+  arr_container.addBool("baz", "baz's description").required(true);
 
   EXPECT_FALSE(inlet.verify());
 }
@@ -198,13 +198,13 @@ TYPED_TEST(inlet_object, simple_array_of_struct_verify_lambda)
   DataStore ds;
   Inlet inlet = createBasicInlet<TypeParam>(&ds, testString);
 
-  auto& arr_table = inlet.addStructArray("foo");
+  auto& arr_container = inlet.addStructArray("foo");
 
-  arr_table.addBool("bar", "bar's description").required();
+  arr_container.addBool("bar", "bar's description").required();
 
   // Ensuring that foo is the element of the collection
-  arr_table.registerVerifier(
-    [](const axom::inlet::Table& foo) { return foo.contains("bar"); });
+  arr_container.registerVerifier(
+    [](const axom::inlet::Container& foo) { return foo.contains("bar"); });
 
   EXPECT_TRUE(inlet.verify());
 }
@@ -217,13 +217,13 @@ TYPED_TEST(inlet_object, simple_array_of_struct_verify_lambda_pass)
   DataStore ds;
   Inlet inlet = createBasicInlet<TypeParam>(&ds, testString);
 
-  auto& arr_table = inlet.addStructArray("foo");
+  auto& arr_container = inlet.addStructArray("foo");
 
-  arr_table.addBool("bar", "bar's description");
-  arr_table.addBool("baz", "baz's description");
+  arr_container.addBool("bar", "bar's description");
+  arr_container.addBool("baz", "baz's description");
 
   // Can specify either "bar" or "baz" but not both
-  arr_table.registerVerifier([](const axom::inlet::Table& foo) {
+  arr_container.registerVerifier([](const axom::inlet::Container& foo) {
     return !(foo.contains("bar") && foo.contains("baz"));
   });
 
@@ -238,13 +238,13 @@ TYPED_TEST(inlet_object, simple_array_of_struct_verify_lambda_fail)
   DataStore ds;
   Inlet inlet = createBasicInlet<TypeParam>(&ds, testString);
 
-  auto& arr_table = inlet.addStructArray("foo");
+  auto& arr_container = inlet.addStructArray("foo");
 
-  arr_table.addBool("bar", "bar's description");
-  arr_table.addBool("baz", "baz's description");
+  arr_container.addBool("bar", "bar's description");
+  arr_container.addBool("baz", "baz's description");
 
   // Can specify either "bar" or "baz" but not both
-  arr_table.registerVerifier([](const axom::inlet::Table& foo) {
+  arr_container.registerVerifier([](const axom::inlet::Container& foo) {
     return !(foo.contains("bar") && foo.contains("baz"));
   });
 
@@ -260,7 +260,7 @@ struct FooWithArray
 template <>
 struct FromInlet<FooWithArray>
 {
-  FooWithArray operator()(const axom::inlet::Table& base)
+  FooWithArray operator()(const axom::inlet::Container& base)
   {
     FooWithArray f = {base["arr"]};
     return f;
@@ -275,9 +275,9 @@ TYPED_TEST(inlet_object, array_of_struct_containing_array)
   DataStore ds;
   Inlet inlet = createBasicInlet<TypeParam>(&ds, testString);
 
-  auto& arr_table = inlet.addStructArray("foo");
+  auto& arr_container = inlet.addStructArray("foo");
 
-  arr_table.addIntArray("arr", "arr's description");
+  arr_container.addIntArray("arr", "arr's description");
   // Contiguous indexing for generality
   std::unordered_map<int, FooWithArray> expected_foos = {{0, {{{0, 3}}}},
                                                          {1, {{{0, 2}}}}};
@@ -299,7 +299,7 @@ struct MoveOnlyFoo
 template <>
 struct FromInlet<MoveOnlyFoo>
 {
-  MoveOnlyFoo operator()(const axom::inlet::Table& base)
+  MoveOnlyFoo operator()(const axom::inlet::Container& base)
   {
     MoveOnlyFoo f(base["bar"], base["baz"]);
     return f;
@@ -459,21 +459,21 @@ TYPED_TEST(inlet_object, composite_type_checks)
   inlet.addIntArray("luaArrays/arr1");
   inlet.addBoolArray("luaArrays/arr2");
 
-  auto arr_table = inlet["luaArrays"];
-  // The table containing the two arrays is not an array, but an object
-  EXPECT_EQ(arr_table.type(), InletType::Object);
+  auto arr_container = inlet["luaArrays"];
+  // The container containing the two arrays is not an array, but an object
+  EXPECT_EQ(arr_container.type(), InletType::Object);
 
   // But the things it contains are arrays
-  EXPECT_EQ(arr_table["arr1"].type(), InletType::Collection);
-  EXPECT_EQ(arr_table["arr2"].type(), InletType::Collection);
+  EXPECT_EQ(arr_container["arr1"].type(), InletType::Collection);
+  EXPECT_EQ(arr_container["arr2"].type(), InletType::Collection);
 
-  auto foo_table = inlet["foo"];
-  // Similarly, the table containing the two bools is an object
-  EXPECT_EQ(foo_table.type(), InletType::Object);
+  auto foo_container = inlet["foo"];
+  // Similarly, the container containing the two bools is an object
+  EXPECT_EQ(foo_container.type(), InletType::Object);
 
   // But the things it contains are booleans
-  EXPECT_EQ(foo_table["bar"].type(), InletType::Bool);
-  EXPECT_EQ(foo_table["baz"].type(), InletType::Bool);
+  EXPECT_EQ(foo_container["bar"].type(), InletType::Bool);
+  EXPECT_EQ(foo_container["baz"].type(), InletType::Bool);
 }
 
 TYPED_TEST(inlet_object, implicit_conversion_primitives)
@@ -532,7 +532,7 @@ struct BarWithFooWithArray
 template <>
 struct FromInlet<BarWithFooWithArray>
 {
-  BarWithFooWithArray operator()(const axom::inlet::Table& base)
+  BarWithFooWithArray operator()(const axom::inlet::Container& base)
   {
     BarWithFooWithArray b;
     b.foo = base["foo"].get<FooWithArray>();
@@ -548,9 +548,9 @@ TYPED_TEST(inlet_object, nested_array_of_struct_containing_array)
   DataStore ds;
   Inlet inlet = createBasicInlet<TypeParam>(&ds, testString);
 
-  auto& bar_table = inlet.addStructArray("bars");
-  auto& foo_table = bar_table.addStruct("foo");
-  foo_table.addIntArray("arr", "arr's description");
+  auto& bar_container = inlet.addStructArray("bars");
+  auto& foo_container = bar_container.addStruct("foo");
+  foo_container.addIntArray("arr", "arr's description");
 
   // Contiguous indexing for generality
   std::unordered_map<int, BarWithFooWithArray> expected_bars = {
@@ -574,7 +574,7 @@ struct QuuxWithFooArray
 template <>
 struct FromInlet<QuuxWithFooArray>
 {
-  QuuxWithFooArray operator()(const axom::inlet::Table& base)
+  QuuxWithFooArray operator()(const axom::inlet::Container& base)
   {
     QuuxWithFooArray q;
     q.arr = base["arr"].get<std::unordered_map<int, Foo>>();
@@ -592,11 +592,11 @@ TYPED_TEST(inlet_object, nested_array_of_struct)
   DataStore ds;
   Inlet inlet = createBasicInlet<TypeParam>(&ds, testString);
 
-  auto& quux_table = inlet.addStructArray("quux");
-  auto& foo_table = quux_table.addStructArray("arr");
+  auto& quux_container = inlet.addStructArray("quux");
+  auto& foo_container = quux_container.addStructArray("arr");
 
-  foo_table.addBool("bar", "bar's description");
-  foo_table.addBool("baz", "baz's description");
+  foo_container.addBool("bar", "bar's description");
+  foo_container.addBool("baz", "baz's description");
 
   // Contiguous indexing for generality
   std::unordered_map<int, QuuxWithFooArray> expected_quuxs = {
@@ -617,11 +617,11 @@ TYPED_TEST(inlet_object, nested_dict_of_array_of_struct)
   DataStore ds;
   Inlet inlet = createBasicInlet<TypeParam>(&ds, testString);
 
-  auto& quux_table = inlet.addStructDictionary("quux");
-  auto& foo_table = quux_table.addStructArray("arr");
+  auto& quux_container = inlet.addStructDictionary("quux");
+  auto& foo_container = quux_container.addStructArray("arr");
 
-  foo_table.addBool("bar", "bar's description");
-  foo_table.addBool("baz", "baz's description");
+  foo_container.addBool("bar", "bar's description");
+  foo_container.addBool("baz", "baz's description");
 
   // Contiguous indexing for generality
   std::unordered_map<std::string, QuuxWithFooArray> expected_quuxs = {
@@ -645,7 +645,7 @@ struct CorgeWithQuuxDictionary
 template <>
 struct FromInlet<CorgeWithQuuxDictionary>
 {
-  CorgeWithQuuxDictionary operator()(const axom::inlet::Table& base)
+  CorgeWithQuuxDictionary operator()(const axom::inlet::Container& base)
   {
     CorgeWithQuuxDictionary c;
     c.dict =
@@ -670,12 +670,12 @@ TYPED_TEST(inlet_object, nested_array_of_dict_of_array_of_struct)
   DataStore ds;
   Inlet inlet = createBasicInlet<TypeParam>(&ds, testString);
 
-  auto& corge_table = inlet.addStructArray("corge");
-  auto& quux_table = corge_table.addStructDictionary("dict");
-  auto& foo_table = quux_table.addStructArray("arr");
+  auto& corge_container = inlet.addStructArray("corge");
+  auto& quux_container = corge_container.addStructDictionary("dict");
+  auto& foo_container = quux_container.addStructArray("arr");
 
-  foo_table.addBool("bar", "bar's description");
-  foo_table.addBool("baz", "baz's description");
+  foo_container.addBool("bar", "bar's description");
+  foo_container.addBool("baz", "baz's description");
 
   // Contiguous indexing for generality
   std::unordered_map<int, CorgeWithQuuxDictionary> expected_corges = {
@@ -703,7 +703,7 @@ struct QuuxWithFooWithArray
 template <>
 struct FromInlet<QuuxWithFooWithArray>
 {
-  QuuxWithFooWithArray operator()(const axom::inlet::Table& base)
+  QuuxWithFooWithArray operator()(const axom::inlet::Container& base)
   {
     QuuxWithFooWithArray q;
     q.arr = base["outer_arr"].get<std::unordered_map<int, FooWithArray>>();
@@ -721,10 +721,10 @@ TYPED_TEST(inlet_object, nested_dict_of_array_of_struct_with_array)
   DataStore ds;
   Inlet inlet = createBasicInlet<TypeParam>(&ds, testString);
 
-  auto& quux_table = inlet.addStructDictionary("quux");
-  auto& foo_table = quux_table.addStructArray("outer_arr");
+  auto& quux_container = inlet.addStructDictionary("quux");
+  auto& foo_container = quux_container.addStructArray("outer_arr");
 
-  foo_table.addIntArray("arr", "arr's description");
+  foo_container.addIntArray("arr", "arr's description");
 
   // Contiguous indexing for generality
   std::unordered_map<std::string, QuuxWithFooWithArray> expected_quuxs = {
@@ -748,7 +748,7 @@ struct QuuxWithSingleFoo
 template <>
 struct FromInlet<QuuxWithSingleFoo>
 {
-  QuuxWithSingleFoo operator()(const axom::inlet::Table& base)
+  QuuxWithSingleFoo operator()(const axom::inlet::Container& base)
   {
     QuuxWithSingleFoo q;
     q.foo = base["foo"].get<Foo>();
@@ -802,6 +802,58 @@ TYPED_TEST(inlet_object, primitive_arrays_as_std_vector)
   EXPECT_EQ(arr_w_indices, expected_arr_w_indices);
 }
 
+TYPED_TEST(inlet_object, primitive_arrays_as_std_vector_wrong_type)
+{
+  std::string testString = " arr = { [0] = 'a', [1] = 'b', [2] = 'c'}";
+  DataStore ds;
+  Inlet inlet = createBasicInlet<TypeParam>(&ds, testString);
+
+  // Define schema
+  inlet.addIntArray("arr");
+
+  // The array was empty (same as if it didn't exist), but *not* marked as required
+  EXPECT_TRUE(inlet.verify());
+
+  // Attempt both construction and assignment
+  std::vector<int> expected_arr {};
+  std::vector<int> arr = inlet["arr"];
+  EXPECT_EQ(arr, expected_arr);
+  arr = inlet["arr"];
+  EXPECT_EQ(arr, expected_arr);
+}
+
+TYPED_TEST(inlet_object, primitive_arrays_as_std_vector_wrong_type_reqd_fail)
+{
+  std::string testString = " arr = { [0] = 'a', [1] = 'b', [2] = 'c'}";
+  DataStore ds;
+  Inlet inlet = createBasicInlet<TypeParam>(&ds, testString);
+
+  // Define schema
+  inlet.addIntArray("arr").required();
+
+  // The array was empty (same as if it didn't exist), but marked as required
+  EXPECT_FALSE(inlet.verify());
+}
+
+TYPED_TEST(inlet_object, primitive_arrays_as_std_vector_mixed_type)
+{
+  std::string testString = " arr = { [0] = 4, [1] = 6, [2] = 'a', [3] = 'b'}";
+  DataStore ds;
+  Inlet inlet = createBasicInlet<TypeParam>(&ds, testString);
+
+  // Define schema
+  inlet.addIntArray("arr");
+
+  EXPECT_TRUE(inlet.verify());
+
+  // Attempt both construction and assignment
+  std::vector<int> expected_arr {4, 6};
+  std::vector<int> arr = inlet["arr"];
+  EXPECT_EQ(arr, expected_arr);
+  arr = inlet["arr"];
+  EXPECT_EQ(arr, expected_arr);
+}
+
 TYPED_TEST(inlet_object, struct_arrays_as_std_vector)
 {
   std::string testString =
@@ -810,10 +862,10 @@ TYPED_TEST(inlet_object, struct_arrays_as_std_vector)
   DataStore ds;
   Inlet inlet = createBasicInlet<TypeParam>(&ds, testString);
 
-  auto& arr_table = inlet.addStructArray("foo");
+  auto& arr_container = inlet.addStructArray("foo");
 
-  arr_table.addBool("bar", "bar's description");
-  arr_table.addBool("baz", "baz's description");
+  arr_container.addBool("bar", "bar's description");
+  arr_container.addBool("baz", "baz's description");
   std::vector<Foo> expected_foos = {{true, false}, {false, true}};
   auto foos = inlet["foo"].get<std::vector<Foo>>();
   EXPECT_EQ(foos, expected_foos);
@@ -847,10 +899,10 @@ TYPED_TEST(inlet_object_dict, simple_dict_of_struct_by_value)
   DataStore ds;
   Inlet inlet = createBasicInlet<TypeParam>(&ds, testString);
 
-  auto& dict_table = inlet.addStructDictionary("foo");
+  auto& dict_container = inlet.addStructDictionary("foo");
 
-  dict_table.addBool("bar", "bar's description");
-  dict_table.addBool("baz", "baz's description");
+  dict_container.addBool("bar", "bar's description");
+  dict_container.addBool("baz", "baz's description");
   std::unordered_map<std::string, Foo> expected_foos = {{"key1", {true, false}},
                                                         {"key2", {false, true}}};
   std::unordered_map<std::string, Foo> foos;
@@ -867,7 +919,7 @@ struct FooWithDict
 template <>
 struct FromInlet<FooWithDict>
 {
-  FooWithDict operator()(const axom::inlet::Table& base)
+  FooWithDict operator()(const axom::inlet::Container& base)
   {
     FooWithDict f = {base["arr"]};
     return f;
@@ -882,9 +934,9 @@ TYPED_TEST(inlet_object_dict, dict_of_struct_containing_dict)
   DataStore ds;
   Inlet inlet = createBasicInlet<TypeParam>(&ds, testString);
 
-  auto& dict_table = inlet.addStructDictionary("foo");
+  auto& dict_container = inlet.addStructDictionary("foo");
 
-  dict_table.addIntDictionary("arr", "arr's description");
+  dict_container.addIntDictionary("arr", "arr's description");
   std::unordered_map<std::string, FooWithDict> expected_foos = {
     {"key3", {{{"key1", 3}}}},
     {"key4", {{{"key2", 2}}}}};
@@ -902,9 +954,9 @@ TYPED_TEST(inlet_object_dict, dict_of_struct_containing_array)
   DataStore ds;
   Inlet inlet = createBasicInlet<TypeParam>(&ds, testString);
 
-  auto& dict_table = inlet.addStructDictionary("foo");
+  auto& dict_container = inlet.addStructDictionary("foo");
 
-  dict_table.addIntArray("arr", "arr's description");
+  dict_container.addIntArray("arr", "arr's description");
   std::unordered_map<std::string, FooWithArray> expected_foos = {
     {"key3", {{{0, 3}}}},
     {"key4", {{{0, 2}}}}};
@@ -922,9 +974,9 @@ TYPED_TEST(inlet_object_dict, array_of_struct_containing_dict)
   DataStore ds;
   Inlet inlet = createBasicInlet<TypeParam>(&ds, testString);
 
-  auto& arr_table = inlet.addStructArray("foo");
+  auto& arr_container = inlet.addStructArray("foo");
 
-  arr_table.addIntDictionary("arr", "arr's description");
+  arr_container.addIntDictionary("arr", "arr's description");
   std::unordered_map<int, FooWithDict> expected_foos = {{0, {{{"key1", 3}}}},
                                                         {1, {{{"key2", 2}}}}};
   std::unordered_map<int, FooWithDict> foos_with_dict;
@@ -975,9 +1027,9 @@ TEST(inlet_object_lua, array_of_struct_containing_array)
   DataStore ds;
   Inlet inlet = createBasicInlet<axom::inlet::LuaReader>(&ds, testString);
 
-  auto& arr_table = inlet.addStructArray("foo");
+  auto& arr_container = inlet.addStructArray("foo");
 
-  arr_table.addIntArray("arr", "arr's description");
+  arr_container.addIntArray("arr", "arr's description");
   std::unordered_map<int, FooWithArray> expected_foos = {{4, {{{1, 3}}}},
                                                          {7, {{{6, 2}}}}};
   std::unordered_map<int, FooWithArray> foos_with_arr;
@@ -1047,10 +1099,10 @@ TEST(inlet_object_lua, struct_arrays_as_std_vector_discontiguous)
   DataStore ds;
   Inlet inlet = createBasicInlet<axom::inlet::LuaReader>(&ds, testString);
 
-  auto& arr_table = inlet.addStructArray("foo");
+  auto& arr_container = inlet.addStructArray("foo");
 
-  arr_table.addBool("bar", "bar's description");
-  arr_table.addBool("baz", "baz's description");
+  arr_container.addBool("bar", "bar's description");
+  arr_container.addBool("baz", "baz's description");
   std::vector<Foo> expected_foos = {{true, false}, {false, true}};
   auto foos = inlet["foo"].get<std::vector<Foo>>();
   EXPECT_EQ(foos, expected_foos);
@@ -1081,10 +1133,10 @@ TEST(inlet_object_lua, struct_arrays_as_std_vector_implicit_idx)
   DataStore ds;
   Inlet inlet = createBasicInlet<axom::inlet::LuaReader>(&ds, testString);
 
-  auto& arr_table = inlet.addStructArray("foo");
+  auto& arr_container = inlet.addStructArray("foo");
 
-  arr_table.addBool("bar", "bar's description");
-  arr_table.addBool("baz", "baz's description");
+  arr_container.addBool("bar", "bar's description");
+  arr_container.addBool("baz", "baz's description");
   std::vector<Foo> expected_foos = {{true, false}, {false, true}};
   auto foos = inlet["foo"].get<std::vector<Foo>>();
   EXPECT_EQ(foos, expected_foos);
@@ -1098,9 +1150,9 @@ TEST(inlet_object_lua_dict, dict_of_struct_containing_array)
   DataStore ds;
   Inlet inlet = createBasicInlet<axom::inlet::LuaReader>(&ds, testString);
 
-  auto& dict_table = inlet.addStructDictionary("foo");
+  auto& dict_container = inlet.addStructDictionary("foo");
 
-  dict_table.addIntArray("arr", "arr's description");
+  dict_container.addIntArray("arr", "arr's description");
   std::unordered_map<std::string, FooWithArray> expected_foos = {
     {"key3", {{{1, 3}}}},
     {"key4", {{{6, 2}}}}};
@@ -1118,9 +1170,9 @@ TEST(inlet_object_lua_dict, array_of_struct_containing_dict)
   DataStore ds;
   Inlet inlet = createBasicInlet<axom::inlet::LuaReader>(&ds, testString);
 
-  auto& arr_table = inlet.addStructArray("foo");
+  auto& arr_container = inlet.addStructArray("foo");
 
-  arr_table.addIntDictionary("arr", "arr's description");
+  arr_container.addIntDictionary("arr", "arr's description");
   std::unordered_map<int, FooWithDict> expected_foos = {{7, {{{"key1", 3}}}},
                                                         {4, {{{"key2", 2}}}}};
   std::unordered_map<int, FooWithDict> foos_with_dict;
@@ -1172,10 +1224,10 @@ TEST(inlet_object_lua_dict, mixed_keys_object)
   DataStore ds;
   Inlet inlet = createBasicInlet<axom::inlet::LuaReader>(&ds, testString);
 
-  auto& dict_table = inlet.addStructDictionary("foo");
+  auto& dict_container = inlet.addStructDictionary("foo");
 
-  dict_table.addBool("bar", "bar's description");
-  dict_table.addBool("baz", "baz's description");
+  dict_container.addBool("bar", "bar's description");
+  dict_container.addBool("baz", "baz's description");
   std::unordered_map<VariantKey, Foo> expected_foos = {{"key1", {true, false}},
                                                        {1, {false, true}}};
   std::unordered_map<VariantKey, Foo> foos;